add "help" option
Run as:

    git get-merge help

Technically we should integrate nicely with git's command line parsing
which knows to look for "--help" and "-h", but one problem at a time.
If you just happen to have a git reference named exactly "help", you'll
have to expand that reference to a sha hash, which will never ever be
"help" (because hexadecimal doesn't allow for "l" or "p").